Fixing layout and danger mode
