+++ src/services/preact-canvas/components/game/style.css
@@ -1,7 +1,7 @@
 .game {
   display: flex;
-  flex-flow: column;
   flex: 1;
+  align-items: center;
 }
 
 .toggle-label {
