+++ src/services/preact-canvas/index.tsx
@@ -106,30 +106,6 @@ class PreactService extends Component<Props, State> {
     );
   }
 
-  componentDidMount() {
-    window.addEventListener("keydown", this._onKeyDown);
-    window.addEventListener("keyup", this._onKeyUp);
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener("keydown", this._onKeyDown);
-    window.removeEventListener("keyup", this._onKeyUp);
-  }
-
-  @bind
-  private _onKeyDown(event: KeyboardEvent) {
-    if (event.key === "Shift") {
-      this._onDangerModeChange(!this.state.dangerMode);
-    }
-  }
-
-  @bind
-  private _onKeyUp(event: KeyboardEvent) {
-    if (event.key === "Shift") {
-      this._onDangerModeChange(!this.state.dangerMode);
-    }
-  }
-
   @bind
   private _onDangerModeChange(dangerMode: boolean) {
     this.setState({ dangerMode });
